Remove strdup, it was causing crashes, use malloc with strcpy instead

+++ include/file_transport.h
@@ -56,7 +56,7 @@ typedef struct file_transport
 }file_transport_t;
 
 file_transport_t * file_transport_new(string_t filename);
-int file_transport_setopt(file_transport_t *transport, transport_option_t option, unsigned long data);
+int file_transport_setopt(file_transport_t *transport, transport_option_t option, unsigned long long int data);
 
 #ifdef __cplusplus
 }
